+++ packages/ts-components/src/components/recommended-articles/RecommendedArticles.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import get from 'lodash.get';
-
-import RelatedArticles from '@times-components/related-articles';
+import {
+  Slice,
+  SliceArticle,
+  MouseEventType
+} from '@times-components/ts-slices';
 
 import { useFetch } from '../../helpers/fetch/FetchProvider';
 import { useTrackingContext } from '../../helpers/tracking/TrackingContextProvider';
-import { getRelatedArticlesSlice } from './formatters';
+import { getRecommendedArticlesSlice } from './formatters';
+
+import { Header } from './styles';
 
 export const RecommendedArticles: React.FC<{
   heading: string;
@@ -24,30 +29,21 @@ export const RecommendedArticles: React.FC<{
 
   const { fireAnalyticsEvent } = useTrackingContext();
 
-  const slice = getRelatedArticlesSlice(articles);
-
-  const onClickHandler = (__: MouseEvent, article: { url: string }) => {
-    const found = slice.items.find(
-      item => item.article.shortIdentifier === article.url.slice(-9)
-    );
-
+  const onClickHandler = (__: MouseEventType, article: SliceArticle) => {
     if (fireAnalyticsEvent) {
       fireAnalyticsEvent({
         action: 'Clicked',
-        attrs: { article_parent_name: found ? found.article.headline : '' }
+        attrs: { article_parent_name: article.headline }
       });
     }
   };
 
   return (
     <div id="recommended-articles">
-      <RelatedArticles
-        heading={heading}
-        slice={slice}
-        isVisible
-        onPress={onClickHandler}
-        // tslint:disable-next-line: no-empty
-        analyticsStream={() => {}}
+      <Header>{heading}</Header>
+      <Slice
+        slice={getRecommendedArticlesSlice(articles)}
+        clickHandler={onClickHandler}
       />
     </div>
   );