+++ packages/ts-components/src/components/recommended-articles/formatters.ts
@@ -1,8 +1,8 @@
-import { RelatedArticleSliceType } from '../../types/related-article-slice';
+import { SliceArticle } from '@times-components/ts-slices';
 
 // TYPES
 
-export type Byline =
+type Byline =
   | { __typename: string; type: 'author'; name: string }
   | { __typename: string; type: 'inline'; value: string };
 
@@ -14,10 +14,10 @@ type MediaCrop = {
 };
 type Media = { __typename: string; crops: MediaCrop[] };
 
-export type SummaryText = { __typename: string; text: string };
+type SummaryText = { __typename: string; text: string };
 type Summary = { __typename: string; children: SummaryText[] };
 
-export type Article = {
+type Article = {
   __typename: string;
   url: string;
   slug: string;
@@ -31,68 +31,58 @@ export type Article = {
 
 // HELPERS
 
-const getBylineAttr = (byline: Byline) => ({
-  value: byline.type === 'author' ? byline.name : byline.value
-});
+const getBylines = (bylines?: Byline[]) =>
+  bylines
+    ? bylines
+        .map(byline => (byline.type === 'author' ? byline.name : byline.value))
+        .join('')
+    : undefined;
 
-const getByline = (byline: Byline) => ({
-  name: byline.type,
-  children: [{ name: 'text', attributes: getBylineAttr(byline), children: [] }]
-});
+const getSummary = (summary?: Summary) =>
+  summary && summary.children
+    ? summary.children.map(child => child.text).join('') + '...'
+    : undefined;
 
-const getBylines = (bylines?: Byline[]) => {
-  if (bylines) {
-    return bylines.map((byline: Byline) => ({ byline: [getByline(byline)] }));
-  }
-  return undefined;
-};
-
-const getSummaryText = (summary: Summary) => {
-  const text = summary.children
-    .map((child: SummaryText) => child.text)
-    .join('');
-  return { value: text.slice(0, text.slice(0, 125).lastIndexOf(' ')) };
-};
-
-const getSummaryParagraph = (summary: Summary) => [
-  { name: 'text', attributes: getSummaryText(summary), children: [] }
-];
-
-const getSummary = (summary?: Summary) => {
-  if (summary && summary.children) {
-    return [{ name: 'paragraph', children: getSummaryParagraph(summary) }];
-  }
-  return [];
-};
-
-const getImage = (media?: Media) => {
-  if (media && media.crops) {
-    const image = media.crops.find(
-      (crop: MediaCrop) => crop.aspectRatio === '16:9'
-    );
-    return image ? { crop169: { url: image.url }, title: image.alt } : null;
-  }
-  return {};
-};
+const getImageCrops = (media?: Media) =>
+  media && media.crops
+    ? media.crops.map(crop => ({ url: crop.url, ratio: crop.aspectRatio }))
+    : [];
 
 // MAIN
 
-export const getRelatedArticlesSlice = (
-  articles: any
-): RelatedArticleSliceType => ({
-  sliceName: 'StandardSlice',
-  items: articles
-    .map((article: Article) => ({
-      article: {
-        slug: article.slug,
-        shortIdentifier: article.url.slice(-9),
-        label: article.label,
-        headline: article.headline,
-        publishedTime: article.publishedDateTime,
-        bylines: getBylines(article.bylines),
-        summary125: getSummary(article.summary),
-        leadAsset: getImage(article.media)
-      }
-    }))
-    .slice(0, 3)
-});
+const formatArticle = (article: Article): SliceArticle => {
+  return {
+    url: article.url,
+    label: article.label,
+    byline: getBylines(article.bylines),
+    headline: article.headline,
+    summary: getSummary(article.summary),
+    datePublished: article.publishedDateTime,
+    images: {
+      alt: article.headline,
+      crops: getImageCrops(article.media)
+    }
+  };
+};
+
+const getSliceName = (numOfArticles: number) => {
+  switch (numOfArticles) {
+    case 1:
+      return 'RELATED_ARTICLE_1';
+    case 2:
+      return 'RELATED_ARTICLE_2';
+    default:
+      return 'RELATED_ARTICLE_3';
+  }
+};
+
+export const getRecommendedArticlesSlice = (articles: Article[]) => {
+  return {
+    name: getSliceName(articles.length),
+    children: articles
+      .map((article: Article) => ({
+        article: formatArticle(article)
+      }))
+      .slice(0, 3)
+  };
+};