+++ packages/ts-components/src/fixtures/preview-data/recommended-articles.ts
@@ -7,7 +7,7 @@ export default {
         slug: 'save-or-splurge-what-experts-spend-their-own-money-on',
         label: null,
         headline: 'Save or splurge: what experts spend their own money on',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'inline',
@@ -54,7 +54,7 @@ export default {
         slug: 'lieutenant-colonel-ian-crooke-obituary',
         label: 'Obituary',
         headline: 'Lieutenant Colonel Ian Crooke',
-        publishedDateTime: '2022-06-06T16:00:00.000Z',
+        publishedDateTime: '2022-07-27T16:00:00.000Z',
         bylines: [],
         summary: {
           children: [
@@ -81,7 +81,7 @@ export default {
         slug: 'is-the-party-over-for-boris-johnson',
         label: 'RED BOX | PATRICK MAGUIRE',
         headline: 'Is the party over for Johnson?',
-        publishedDateTime: '2022-06-07T08:00:00.000Z',
+        publishedDateTime: '2022-07-27T08:00:00.000Z',
         bylines: [
           {
             type: 'author',
@@ -121,7 +121,7 @@ export default {
         label: 'Political Sketch',
         headline:
           'A PM, his party and the country left firmly in limp-on territory',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'author',
@@ -153,7 +153,7 @@ export default {
         slug: 'boris-johnson-should-look-for-an-honourable-exit',
         label: 'WILLIAM HAGUE | COMMENT',
         headline: 'Hague: Johnson should look for an honourable exit',
-        publishedDateTime: '2022-06-06T21:00:00.000Z',
+        publishedDateTime: '2022-07-26T21:00:00.000Z',
         bylines: [
           {
             type: 'author',
@@ -185,7 +185,7 @@ export default {
         slug: 'carrie-johnson-and-hard-times-at-no-10',
         label: 'POLITICS',
         headline: 'Carrie Johnson and hard times at No 10',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'author',
@@ -218,7 +218,7 @@ export default {
           'dead-man-walking-times-commentators-give-their-verdicts-boris-johnson-confidence-vote',
         label: 'Politics',
         headline: 'Dead man walking? Times commentators give their verdicts',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'inline',
@@ -270,7 +270,7 @@ export default {
         slug: 'margot-heuman-obituary',
         label: 'Obituary',
         headline: 'Margot Heuman',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [],
         summary: {
           children: [
@@ -298,7 +298,7 @@ export default {
           'the-times-view-on-tory-confidence-in-boris-johnson-pyrrhic-victory',
         label: 'LEADING ARTICLE',
         headline: 'Pyrrhic Victory',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'author',
@@ -331,7 +331,7 @@ export default {
           'boris-johnson-allies-took-time-saving-big-dog-leadership-contest',
         label: 'Politics',
         headline: 'Allies took their time in campaign to save Big Dog',
-        publishedDateTime: '2022-06-06T23:01:00.000Z',
+        publishedDateTime: '2022-07-26T23:01:00.000Z',
         bylines: [
           {
             type: 'author',