+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.92](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.7.91...@times-components/article-paragraph@1.7.92) (2022-07-29)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.7.91](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.7.90...@times-components/article-paragraph@1.7.91) (2022-07-28)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.7.90](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.7.89...@times-components/article-paragraph@1.7.90) (2022-07-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.7.89](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.7.88...@times-components/article-paragraph@1.7.89) (2022-07-26)
 
 **Note:** Version bump only for package @times-components/article-paragraph